dodanie zmiennej marzy i wiboru

+++ MortgageCalculator/src/main/java/MortgageCalculator/MortgageCalculator/rest/MortgageRestController.java
@@ -20,7 +20,7 @@ public class MortgageRestController {
 	private Calculator calculator;
 	
 	@GetMapping("/calculate")
-	public List<Result> calculate(@RequestParam double kwota, @RequestParam double okres,
+	public List<Result> calculate(@RequestParam double kwota, @RequestParam int okres,
 			@RequestParam double marza, @RequestParam double wibor, @RequestParam double prowizja, @RequestParam double nadplata,
 			@RequestParam int opoznienieNadplaty, @RequestParam CzestotliwoscNadplat czestotliwoscNadplat) {
 		
@@ -39,7 +39,7 @@ public class MortgageRestController {
 	}
 	
 	@GetMapping("/calculate2")
-	public List<Result> calculate(@RequestParam double kwota, @RequestParam double okres,
+	public List<Result> calculate(@RequestParam double kwota, @RequestParam int okres,
 			@RequestParam double marza, @RequestParam double wibor, @RequestParam double prowizja) {
 		
 		Request request = RequestBuilder.with()//
